requirements.txt
langchain>=0.0.306
openai>=0.27.8 
chromadb>=0.3.29
tiktoken>=0.4.0
pydantic>=1.10.7
litellm>=0.1.819
ipykernel>=6.25.2
python-dotenv~=1.0.0
uvicorn~=0.23.2
fastapi~=0.104.0
gradio~=3.50.2
requests~=2.31.0
pytest~=7.4.0
setuptools~=68.0.0
transitions~=0.9.0